Do not create directory in HDFS when creating ImmutableTablesFulfiller
When creating ImmutableTablesFulfiller, it creates TableManagers, and so
creates HiveTableManager, which then creates
DefaultHdfsDataSourceWriter. This needs RevisionStorage to be created
first. In order to create RevisionStorage a DefaultRevisionProvider,
creates a directory in HDFS to decide which RevisionStorage it should
create.

Above causes a problem when running tests which does not require hdfs to
be provisioned.

By using Provider explicitly, RevisionStorage is created just before any
write or read to or from HDFS,

Pull request: https://github.com/prestodb/tempto/pull/177
